requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.4
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
Faker==15.3.4
flake8==6.0.0
Markdown==3.4.1
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
